api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["aa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.109.2"}
fastapi-cache2 = "^0.2.1"
celery = "^5.3.6"
alembic = "^1.13.1"
redis = "^5.0.1"
gunicorn = "^21.2.0"
watchdog = "^4.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.26"}
asyncpg = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
scikit-learn = "^1.4.1.post1"
werkzeug = "^3.0.1"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
torch = "^2.2.2"
torchvision = "^0.17.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
